refactor: bookmark에도 usecase, adapter, port 적용 (#5)
