+++ report-service/src/main/java/com/songspasssta/reportservice/application/port/in/BookmarkUseCase.java
@@ -1,0 +1,10 @@
+package com.songspasssta.reportservice.application.port.in;
+
+import com.songspasssta.reportservice.dto.response.BookmarkedReportsResponse;
+
+public interface BookmarkUseCase {
+    BookmarkedReportsResponse findMyBookmarks(Long memberId);
+
+    String toggleBookmark(Long reportId, Long memberId);
+}
+
